+++ proxyprovider/filter.go
@@ -3,6 +3,7 @@ package proxyprovider
 import (
 	"strings"
 
+	C "github.com/sagernet/sing-box/constant"
 	"github.com/sagernet/sing-box/option"
 	E "github.com/sagernet/sing/common/exceptions"
 
@@ -108,4 +109,64 @@ func (i *FilterItem) match(outbound *option.Outbound, tagMap map[string]string) 
 	}
 	b, err := i.regex.MatchString(s)
 	return err == nil && b
+}
+
+func getServer(outbound *option.Outbound) string {
+	var server string
+	switch outbound.Type {
+	case C.TypeHTTP:
+		server = outbound.HTTPOptions.Server
+	case C.TypeShadowsocks:
+		server = outbound.ShadowsocksOptions.Server
+	case C.TypeVMess:
+		server = outbound.VMessOptions.Server
+	case C.TypeTrojan:
+		server = outbound.TrojanOptions.Server
+	case C.TypeWireGuard:
+		server = outbound.WireGuardOptions.Server
+	case C.TypeHysteria:
+		server = outbound.HysteriaOptions.Server
+	case C.TypeSSH:
+		server = outbound.SSHOptions.Server
+	case C.TypeShadowTLS:
+		server = outbound.ShadowTLSOptions.Server
+	case C.TypeShadowsocksR:
+		server = outbound.ShadowsocksROptions.Server
+	case C.TypeVLESS:
+		server = outbound.VLESSOptions.Server
+	case C.TypeTUIC:
+		server = outbound.TUICOptions.Server
+	case C.TypeHysteria2:
+		server = outbound.Hysteria2Options.Server
+	}
+	return server
+}
+
+func setServer(outbound *option.Outbound, server string) {
+	switch outbound.Type {
+	case C.TypeHTTP:
+		outbound.HTTPOptions.Server = server
+	case C.TypeShadowsocks:
+		outbound.ShadowsocksOptions.Server = server
+	case C.TypeVMess:
+		outbound.VMessOptions.Server = server
+	case C.TypeTrojan:
+		outbound.TrojanOptions.Server = server
+	case C.TypeWireGuard:
+		outbound.WireGuardOptions.Server = server
+	case C.TypeHysteria:
+		outbound.HysteriaOptions.Server = server
+	case C.TypeSSH:
+		outbound.SSHOptions.Server = server
+	case C.TypeShadowTLS:
+		outbound.ShadowTLSOptions.Server = server
+	case C.TypeShadowsocksR:
+		outbound.ShadowsocksROptions.Server = server
+	case C.TypeVLESS:
+		outbound.VLESSOptions.Server = server
+	case C.TypeTUIC:
+		outbound.TUICOptions.Server = server
+	case C.TypeHysteria2:
+		outbound.Hysteria2Options.Server = server
+	}
 }
